pyproject.toml
[tool.poetry]
name = "pingslurp"
version = "0.1.0"
description = "Keeping up with the Podpings"
authors = ["Brian of London <brian@podping.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lighthive = "^0.4.0"
pydantic = "^1.10.2"
black = "^22.10.0"
asgiref = "^3.5.2"
single-source = "^0.3.0"
motor = "^3.1.1"
beem = {git = "https://gitlab.syncad.com/hive/beem.git"}
httpx = "^0.23.0"
pymssql = "^2.2.5"
python-dotenv = "^0.21.0"
typer = {extras = ["all"], version = "^0.6.1"}


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"